requirements.txt
# Core Dependencies
discord.py>=2.3.2      # Discord bot framework
python-dotenv>=1.0.0   # Environment variable management

# Logging
colorlog>=6.7.0        # Colored console logging output

# Async Support
aiohttp>=3.9.1         # Async HTTP client (required by discord.py)

# Development & Testing (Optional)
pytest>=7.4.3          # Testing framework
flake8>=6.1.0          # Code linting
black>=23.11.0         # Code formatting